requirements.txt
python-dotenv==1.0.1
elasticsearch==8.15.1
